Availability Manager for a Data Center
Using the VI API to simulate a fail over in a virtual machine present
in the data centers and handling it using cloning and migrating the VM
over to another host. VShpere is used to communicated with the ESXi
server.
Successfully migrated the VM to another host with the help of snapshots
(VMDK image file). Alarms were set to inform if a VM was powered off.
Periodically pinging the VM to check if the VM is alive or dead.